more docs and refactor
