+++ src/microsim/allen/__init__.py
@@ -1,3 +1,15 @@
-from .fetch import ApiCellTypesSpecimenDetail, Specimen, get_reconstructions
+from ._fetch import (
+    ApiCellTypesSpecimenDetail,
+    NeuronReconstruction,
+    Specimen,
+    get_reconstructions,
+)
+from ._swc import SWC
 
-__all__ = ["get_reconstructions", "Specimen", "ApiCellTypesSpecimenDetail"]
+__all__ = [
+    "ApiCellTypesSpecimenDetail",
+    "get_reconstructions",
+    "NeuronReconstruction",
+    "Specimen",
+    "SWC",
+]
